refactor: useQuestionList 훅으로 분리
